requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI & ML
openai==1.3.0
anthropic==0.7.0
langchain==0.1.0
chromadb==0.4.18

# Task queue & Background jobs
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
sentry-sdk==1.39.1
prometheus-client==0.19.0
loguru==0.7.2

# Payments
stripe==7.8.0

# Email
sendgrid==6.11.0
jinja2==3.1.2

# HTTP & Requests
httpx==0.25.2
aiohttp==3.9.1

# Utils
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
